test(entire): preserve exact base shard topology
X-Agent: codex/6291-entire-fleet-capture
